+++ src/i18n/enUS/index.ts
@@ -8,6 +8,14 @@ export const blogTranslation = {
   seeAllPosts: 'All posts',
 };
 
+export const socialMediaSharingArticleTranslation = {
+  copy: 'Copy',
+  copied: 'Copied',
+  copiedDescription: 'Link copied to clipboard',
+  share: 'Share',
+  linkToThePost: 'Link to the post',
+};
+
 export const navbarTranslation = {
   home: 'Home',
   about: 'About',